+++ src/main/java/ru/vyarus/guice/persist/orient/support/repository/mixin/crud/delegate/DocumentCrudDelegate.java
@@ -1,4 +1,4 @@
-package ru.vyarus.guice.persist.orient.support.repository.mixin.crud;
+package ru.vyarus.guice.persist.orient.support.repository.mixin.crud.delegate;
 
 import com.google.inject.ProvidedBy;
 import com.google.inject.internal.DynamicSingletonProvider;
@@ -7,6 +7,7 @@ import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import ru.vyarus.guice.persist.orient.repository.delegate.ext.generic.Generic;
+import ru.vyarus.guice.persist.orient.support.repository.mixin.crud.DocumentCrud;
 
 import javax.inject.Inject;
 import javax.inject.Provider;