+++ src/main/java/ru/vyarus/guice/persist/orient/support/repository/mixin/graph/ObjectVertexCrud.java
@@ -1,0 +1,39 @@
+package ru.vyarus.guice.persist.orient.support.repository.mixin.graph;
+
+import com.orientechnologies.orient.core.id.ORID;
+import ru.vyarus.guice.persist.orient.repository.delegate.Delegate;
+import ru.vyarus.guice.persist.orient.support.repository.mixin.crud.BaseObjectCrud;
+import ru.vyarus.guice.persist.orient.support.repository.mixin.graph.delegate.ObjectVertexCrudDelegate;
+
+/**
+ * Crud mixin for object repositories.
+ * Note that {@link ru.vyarus.guice.persist.orient.support.repository.mixin.crud.ObjectCrud}
+ * is used for pure objects and this crud is intended to be used for graph nodes (vertex).
+ * Most likely such objects are annotated with
+ * {@link ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.type.vertex.VertexType}
+ * (or simply extend V in general).
+ * <p>It is important to use this mixin for vertex types, because delete methods use graph api, which
+ * grants consistency checks.</p>
+ *
+ * @param <T> entity type
+ * @author Vyacheslav Rusakov
+ * @since 12.06.2015
+ */
+@Delegate(ObjectVertexCrudDelegate.class)
+public interface ObjectVertexCrud<T> extends BaseObjectCrud<T> {
+
+    /**
+     * @param entity entity to remove
+     */
+    void delete(T entity);
+
+    /**
+     * @param id entity id to remove
+     */
+    void delete(String id);
+
+    /**
+     * @param id entity id to remove
+     */
+    void delete(ORID id);
+}